pyproject.toml
[tool.poetry]
name = "testtask-itworkin"
version = "0.1.0"
description = ""
authors = ["sunny <stratdn@mail.ru>"]
readme = "README.md"
packages = [{include = "testtask_itworkin"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.19"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.0"}
asyncpg = "^0.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
aiofiles = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
